Add control logic

+++ nonlinear/Pi/BA/0.05/512/organize.sh
@@ -1,19 +1,66 @@
 #!/bin/bash
 
-echo "***START ORGANIZATION***"
-mkdir athdf
-mkdir core
-mkdir dat
-mkdir output
-mkdir rst
-mkdir video
-mkdir xdmf
-mv -v *.athdf athdf/
-mv -v *.dat dat/
-mv -v *.xdmf xdmf/
-mv -v SI.0*.rst rst/
-cp -v SI.final.rst rst/
+echo "Organizing..."
+
+if [[ -f *.athdf ]]
+then
+    if [[ ! -d athdf ]]
+    then
+        mkdir athdf
+    fi
+    mv -v *.athdf athdf/
+fi
+
+if [[ -f core.* ]]
+then
+    if [[ ! -d core ]]
+    then
+        mkdir core
+    fi
+    mv -v core.* core/ 
+fi
+
+if [[ -f *.dat ]]
+then
+    if [[ ! -d dat ]]
+    then
+        mkdir dat
+    fi
+    mv -v *.dat dat/
+fi
+
+if [[ ! -d output ]]
+then
+    mkdir output
+fi
 cp -v SI.hst output/
 mv -v *.o* output/
-mv -v core.* core/
-echo "***END ORGANIZATION***"
+
+if [[ -f *.rst ]]
+then
+    if [[ ! -d rst ]]
+    then
+        mkdir rst
+    fi
+    mv -v 0*.rst rst/
+    if [[ -f *.final.rst ]]
+    then
+        cp -v *.final.rst rst/
+    fi
+fi
+
+if [[ ! -d video ]]
+then
+    mkdir video
+fi
+
+if [[ -f *.xdmf]]
+then
+    if [[ ! -d xdmf ]]
+    then
+        mkdir xdmf
+    fi
+    mv -v *.xdmf xdmf/
+fi
+
+echo "... Done."
